externalize job executor from job itself
