+++ api/batch/src/main/java/it/gov/pagopa/rtd/transaction_filter/batch/BatchExecutor.java
@@ -1,0 +1,77 @@
+package it.gov.pagopa.rtd.transaction_filter.batch;
+
+
+import it.gov.pagopa.rtd.transaction_filter.batch.step.PanReaderStep;
+import it.gov.pagopa.rtd.transaction_filter.batch.step.TransactionFilterStep;
+import java.util.Date;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.JobExecution;
+import org.springframework.batch.core.JobParametersBuilder;
+import org.springframework.batch.core.launch.JobLauncher;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
+
+@Slf4j
+@RequiredArgsConstructor
+public class BatchExecutor {
+
+    private final Job job;
+    private final JobLauncher jobLauncher;
+    private final TransactionFilterStep transactionFilterStep;
+    private final PanReaderStep panReaderStep;
+    private final PathMatchingResourcePatternResolver resolver = new PathMatchingResourcePatternResolver();
+
+
+    /**
+     *
+     * @return Method to start the execution of the transaction filter job
+     * @param startDate starting date for the batch job execution
+     */
+    @SneakyThrows
+    public JobExecution execute(Date startDate) {
+        Resource[] transactionResources = resolver.getResources(transactionFilterStep.getTransactionDirectoryPath() + "/*.csv");
+        transactionResources = TransactionFilterStep.filterValidFilenames(transactionResources);
+
+        String hpanPath = panReaderStep.getHpanDirectoryPath();
+        Resource[] hpanResources = resolver.getResources(hpanPath);
+
+        JobExecution execution = null;
+
+        /*
+          The jobLauncher run method is called only if, based on the configured properties, a matching transaction
+          resource is found, and either the remote pan list recovery is enabled, or a pan list file is available locally
+          on the configured path
+         */
+        if (transactionResources.length > 0 &&
+                (getHpanListRecoveryEnabled() || hpanResources.length>0)) {
+
+            log.info("Found {} {}. Starting filtering process",
+                    transactionResources.length, (transactionResources.length > 1 ? "resources" : "resource")
+            );
+
+            execution = jobLauncher.run(job,
+                    new JobParametersBuilder()
+                            .addDate("startDateTime", startDate)
+                            .toJobParameters());
+            clearStoreService();
+
+        } else {
+            if (transactionResources.length == 0) {
+                log.info("No transaction file has been found on configured path: {}", transactionFilterStep.getTransactionDirectoryPath());
+            }
+            if (!getHpanListRecoveryEnabled() && hpanResources.length==0) {
+                log.info("No hpan file has been found on configured path: {}", hpanPath);
+            }
+        }
+
+        return execution;
+    }
+
+    public void clearStoreService() {
+        storeService.clearAll();
+    }
+
+}
